+++ indra/newview/llstartup.cpp
@@ -350,7 +350,7 @@ bool idle_startup()
 
 	const std::string delims (" ");
 	std::string system;
-	int begIdx, endIdx;
+	size_t begIdx, endIdx;  // <alchemy/>
 	std::string osString = LLAppViewer::instance()->getOSInfo().getOSStringSimple();
 
 	begIdx = osString.find_first_not_of (delims);
